# Conflicts:
#	source/CameraRay.cpp
#	source/GLWindowSDL.h
#	source/InterfacesDB.h
#	source/MainContext.cpp
#	source/Particle.cpp
#	source/Particle.h
#	source/ParticleRender.cpp
#	source/ParticleSystem.cpp
#	source/ParticleSystem.h
#	source/Texture.cpp

+++ source/Clock.h
@@ -10,15 +10,15 @@ protected:
 	bool paused = false;
 
 public:
-	void start();
-	int timeEllapsedMsc();
-	int newFrame();
-	int timeEllapsedLastFrameMsc();
-	void restart();
-	bool isActive();
-	void pause();
-	void goOn();
-	bool isPaused() { return paused; }
+	void		start();
+	int64_t		timeEllapsedMsc();
+	int			newFrame();
+	int			timeEllapsedLastFrameMsc();
+	void		restart();
+	bool		isActive();
+	void		pause();
+	void		goOn();
+	bool		isPaused() { return paused; }
 };
 
 #endif
